# assets/Example-Target-Access-Control-swagger.yaml
openapi: 3.0.4
info:
  title: HelloID Access Control target system example API
  description: 'This example API specifies example calls for developing a new API for an Access Control System that will be used for user provisioning from HelloID.<br>Our Starting point, we grant permissions to user accounts and the permissions assigned to a user will be applied to the accessKeys assigned to that user.</br><br>See the Github repo linked below to download the source code.</br>'
  contact:
    name: Tools4everBV Github
    url: https://github.com/orgs/Tools4everBV/
  version: '1.0'
paths:
  /api/AccessKeyAssignments:
    get:
      tags:
        - AccessKeyAssignments
      summary: Get AccessKeyAssignments for all users (optional)
      description: "<h2>Implementation notes</h2>\r\n            This can be userful for reporting, but is generally not required for HelloID because the Accesskeys are typically not assigned by HelloId"
      operationId: GetAllAccessKeyAssignments
      responses:
        '200':
          description: Returns the list of all AccessKeyAssignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessKeyAssignment'
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - AccessKeyAssignments
      summary: Add AccessKeyAssignment (optional test)
      description: "<h2>Implementation notes</h2>\r\n            We will use this action only for testing this example API, generally we do not assign access keys to users in helloID \r\n            \r\n</br>\r\n            Example:\r\n              \r\n                POST /accesskeyassignments\r\n                {\r\n                   \"accessKeyId\": 1,\r\n                   \"userId\": 1\r\n                }"
      operationId: AddAccessKeyAssignment
      requestBody:
        description: The authorization that will be added.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessKeyAssignment'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKeyAssignment'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - AccessKeyAssignments
      summary: Delete AccessKeyAssignment
      description: "<h2>Implementation notes</h2>\r\n            We will use this action when an AccessKey is revoked from a user. This action does not require a response. A [204 No Content] is sufficient."
      operationId: DeleteAccessKeyAssignmentByValue
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
            format: int32
        - name: accessKeyId
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: ''
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/AccessKeyAssignments/{userId}/Accesskeys':
    get:
      tags:
        - AccessKeyAssignments
      summary: Get all assigned Access keys for a specific user
      description: "<h2>Implementation notes</h2>\r\n            This returns all Access Keys currently assigned to the given user.\r\n            \r\n</br>\r\n            Useful when validating which permissions/roles the user has at a given moment."
      operationId: GetAccessKeysByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Returns a list of AccessKeyAssignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessKey'
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: If the user does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/AccessKeyAssignments/{id}':
    delete:
      tags:
        - AccessKeyAssignments
      summary: Delete AccessKeyAssignment (by Id)
      description: "<h2>Implementation notes</h2>\r\n            We will use this action when an AccessKey is revoked from a user. This action does not require a response. A [204 No Content] is sufficient."
      operationId: DeleteAccessKeyAssignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: The authorization was successfully removed.
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No authorization with the specified Id was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/AccessKeys:
    get:
      tags:
        - AccessKeys
      summary: Get all Accesskeys (optional test)
      description: "<h2>Implementation notes</h2>\r\n            We will use this action only for testing this example API"
      operationId: GetAllAccessKeys
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessKey'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - AccessKeys
      summary: Add Accesskey (optional test)
      description: "<h2>Implementation notes</h2>\r\n            We will use this action only for testing this example API, generally we do not create access keys with HelloID \r\n            \r\n</br>\r\n            Example:\r\n              \r\n                POST /Accesskeys\r\n                {\r\n                   \"displayName\": \"Card001\",\r\n                   \"Type\":  \"Employee, Visitor or temporary\"\r\n                   \"isActive\": \"true\"\r\n                }"
      operationId: AddAccessKey
      requestBody:
        description: The authorization that will be added.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessKey'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKey'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/AccessKeys/{id}':
    patch:
      tags:
        - AccessKeys
      summary: Update AccessKey by Id
      description: "<h2>Implementation notes</h2>\r\n              We will use this action to update an AccessKey. This action does not require a response. A [204 No Content] is sufficient.\r\n             \r\n             Example:\r\n               \r\n                 PATCH /Users/:id\r\n                 [  \r\n                     {\r\n                         \"op\": \"replace\",\r\n                         \"path\": \"isActive\",\r\n                         \"value\": \"true\"\r\n                     }\r\n                 ]"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Operation'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKey'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: ''
...
components:
  schemas:
    AccessKey:
      type: object
      properties:
        id:
          type: integer
          description: "This is the internal / database Id.\r\n\r\n\r\nTypically this value will be set by the application itself.\r\n</br>"
          format: int32
          readOnly: true
        displayName:
          type: string
          description: The DisplayName of the AccessKey
          nullable: true
          example: Card001
        type:
          type: string
          description: The Type of the AccessKey
          nullable: true
          example: Employee
        isActive:
          type: boolean
          description: The state of the AccessKey
          nullable: true
          example: true
      additionalProperties: false
      description: "The AccessKey that are available in the target system. Like; 'Card001' or 'Card002'.\r\nThe Permissions on the accessKey are managed due to the Roles on the user object. The permissions assigned to the user object, will be applied to the assigned Accesskeys."
    AccessKeyAssignment:
      required:
        - accessKeyId
        - userId
      type: object
      properties:
        id:
          type: integer
          description: "This is the internal / database Id.\r\n\r\n\r\nTypically this value will be set by the application itself.\r\n</br>"
          format: int32
          readOnly: true
        accessKeyId:
          type: integer
          description: The Id of the AccessKeyId that will be assigned.
          format: int32
          example: 1
        userId:
          type: integer
          description: The Id of the user who will be assigned to this accessKey.
          format: int32
          example: 1
      additionalProperties: false
      description: "A user can have multiple accessKeys. Each AccessKeyAssignment will only have one accessKey like; Card001 or LicensePlate001\r\n\r\n\r\nAn accessKey will be assigned (granted) to a user.\r\n</br>"
...
    Operation:
      type: object
      properties:
        operationType:
          $ref: '#/components/schemas/OperationType'
        path:
          type: string
          nullable: true
        op:
          type: string
          nullable: true
        from:
          type: string
          nullable: true
        value:
          nullable: true
      additionalProperties: false
    OperationType:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
      type: integer
      format: int32
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    Bearer:
      type: http
      description: The bearer token
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]